+++ monitors.py
@@ -66,7 +66,6 @@ class EventMonitor(BaseMonitor):
     def run(self, env, city: City):
         while True:
             self.data = city.events
-
             if self.chunk_size and len(self.data) > self.chunk_size:
                 self.data = city.pull_events()
                 self.dump()
